Use google search API for '-g' option.
The google search soup parsing appeared to be broken, thus I replaced that with
the deprecated (limited queries per day) json-search API.

Also updated setup.py to be a non-DOS file and to only require simplejson if
json could not be imported. Finally bumped up the beautifulsoup requirement as
setup.py failed with the older version.